requirements.txt
telethon>=1.39.0
PyYaml>=6.0.2